refactor(cli/summary): split command into args/json/human submodules
commands/summary.rs was the largest CLI file (1515 lines), bundling three
presenter concerns. Split into a summary/ module:

- summary/mod.rs (611) — orchestration: SummaryArgs, run, run_inner (dispatch +
  flag-exclusivity validation), arg parsing, ingest, and the test block.
- summary/json.rs (170) — JSON serialization (emit_json, grouped_json_value,
  the *_to_json helpers).
- summary/human.rs (778) — human/table rendering (emit_human, all render_*,
  format_*, coverage cells).

Public surface (SummaryArgs, run) unchanged at crate::commands::summary::*.
Pure code-movement: 47/47 item-set parity, moved bodies byte-identical, all
CLI tests pass, clippy clean, live summary/--by-tool OK.
